It works!!!  Well, sorta...
Fixed two bugs in the ink mixing routine,  that mostly works now, but, it's
far from optimal.  Fixed the use of an unitialized variable in merge_line.
Fixed a bug in relative vertical motion command interpretation.  = -> +=
Changed the behavior on unexpected EOF to print out what was read in so
far, rather than just quitting.

Also removed some excessively verbose debugging statements.

The resulting images do not look good.  There's a line length mismatch problem
somewhere which makes the images look like this:

+------+
|\     +
| \    +
+------+

The lower left triangle contains the right edge of the original image.

The image is also cut short by the unexpected EOF problem.

There's still a lot of work left to do, but, this is quite promising.