Change unprint behavior in response to ESC @ initialization command.  If this is found after raster data has been sent, eject the page and quit.
What was happening was that a trailing ESC @ occuring at the end of some
files was resetting the resolution before writing out the data, which was
a bad thing.

Ejecting the page on this command isn't a perfect simulation if ESC @
occurs in the middle of the file, but, I'll deal with that later.