Change the way command line arguments are parsed to make it easier to add option flags.
Add "-u" option flag.  -u does two things.  It sets the nozzle separation to
1, and it treats all vertical head motion commands as if the argument was
to move down pstate.nozzles ticks.  For Microweave code, this most likely
will have no effect, but for softweave, this essentially "unweaves" the
image so that data from each pass appears separately.  This is not so
useful for image viewing, but can be useful for debugging softweave stuff.