Fix the damn vertical artifacts!  This was another case of sloppy arithmetic.  The last pixel is the first pixel + (n-1) pixels, not n pixels, where n is the number of pixels in the line.
Also fix a bug in the ESC r color selection (missing break).

Also fix a bug in the short form of the set unit command.  (horizontal
relative units were not being set.)

Finally, add a byte counter so that errors can be referenced by their
position in the file.  This is useful for correlating errors with
output from parse-escp2.

This is worth a version 0.9 revision number.