Fix some global replacements which had changed "redistribute" into "cyanistribute".
Remove stuff in the GTK ui conditionalised on "#ifndef GIMP_1_0".
Reasoning is as follows.  The GTK ui is there to support gimp 1.0; the
gimpui ui is preferred for gimp 1.1/1.2.  Therefore, the only reason to
use the gtk ui with GIMP_1_0 not defined is for the developers to test
it, and in those circumstances we should see as near as possible the
same as the gimp 1.0 users see.

Amend code which decides whether a given resolution is valid for escp2
printers to use criteria consistent with the current weave code.