Fixed international A sizes (wrong values with width and height)
Changes printers array so it isn't static (needed for genppd, and
it doesn't matter for the GS driver or GIMP plug-in)

Updated genppd to generate PPD files from the driver list.

Added media size, source, and type to raster driver.

(hopefully this will be all that is left to make it all work)