Slightly change the way Ghostscript is packaged for Debian.
Originally, this patch would allow you to compile a package named
"gs-stp", which would both conflict with and provide the original gs package.

Unfortunately, gs-pdfencrypt has a versioned depends on gs, and dpkg does
not support a versioned provides, so you can't install gs-pdfencrypt with
gs-stp.  At least, not as far as I can figure.

So the solution is to not make a "gs-stp" package, but a new gs package,
with "stp" appended to the version number (5.10stp).  This allows
everything to work.