Remove check for random function that is no longer used.
Major cleanup of dither code:

* Put all of the working data into channels rather than ad hoc CMYK
  variables.

* Generalize the dither matrix stuff a bit.

* Split out dither matrices into their own (new) source file.

Add support for Epson Stylus Pro printers (untested!).

Speed up genppd significantly.

TEMPORARILY remove check for missing drivers in foomatic install.
This needs to be restored when we get ID's for the Stylus Pros.

Finally fix a thorny bug in the GUI that caused options to sometimes
be ignored and may have caused crashes (the problem was memory being
freed out from under something due to a GUI callback).

Gray out the paper size entry boxes when they are inactive (when
standard page sizes are in use).

Add paper size entries for roll papers.

Other miscellaneous cleanup of the GUI.

Renumber the channels so that black is guaranteed to be channel 0.
This simplifies a lot of stuff in the dither code.

Check that parameters have actually been verified at print time.

Fix testdither program.