Fix a large number of logic problems related to stp_vars_t.
Many of these are in the color code; there were a lot of problems with
overriding options and such.  There is a new call stp_prune_inactive
to remove inactive and unused settings, to prevent other code from
tripping all over these settings.

Also, the UI now sets all parameters that it will not display to be
inactive.