New channel architecture.
Splitting into dark/light inks, quadtones, etc. is now done in a new
"channel" layer, between the color code and the dither code.  It
is envisioned that eventually there may be a string of channel
operations, so that the color code would itself be a channel
operation.

This has the desirable effect also of simplifying the dither code,
as the dither code does not have to worry about a black channel that
must be treated specially.  It will make it easier to implement
ink limiting, sliding conversions between dark and light inks based
on black level, etc.

This has been tested to the level prior to this commit.  The fast
dither is behaving badly in 6-colors, but this was determined to
predate this change.

Somewhat to my surprise, quality actually seems to be a bit higher
than it was before.  Perhaps simpler really is better...