dyesub:  Initial multi-cut implementation, #ifdef'd out.
This relies on creating multiple, otherwise-duplicate paper types to
handle the various multicut options.  These duplicate papers have not
(nor will they ever) be committed.

I am committing this now because it has all of the multicut support
options enumerated for the DNP DS40/DS80 and the Shinko S2145 printers,
and the eventual "proper" solution will need this knowledge encoded.