backends: dnpds40:  Respect rules for determining how many print buffers we need for the job.
The printer has two buffers, but some image sizes/parameters require
both buffers for a single image.  When submitting jobs to the printer,
we need to wait until the printer has sufficient free buffers.

The safest approach is to always wait for the printer to be idle, but
this prevents us from pipelining jobs to speed up printing.

Now all of the rules are implemented:

 * Switching between matte<->glossy requires both buffers free
 * all 300dpi only need one
 * with DS80, only need 2 buffers for some 600dpi sizes in matte mode
 * with DS40, always need 2 buffers for > 5x7 prints, and 5x7 in matte.